Fix x86_64 Linux build for sipXmediaAdapterLib tests.
git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@10944 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXmediaAdapterLib/sipXmediaMediaProcessing/src/test/CpPhoneMediaInterfaceTest.cpp
@@ -2121,7 +2121,7 @@ class CpPhoneMediaInterfaceTest : public CppUnit::TestCase
         RTL_EVENT("Prompt", 1);
         printf("playing prompt in source 1\n");
         int waveHeaderSize = 116;
-        CPPUNIT_ASSERT_EQUAL((unsigned)160116, sizeof(sine_330hz_16b_8k_signed));
+        CPPUNIT_ASSERT_EQUAL((size_t)160116, sizeof(sine_330hz_16b_8k_signed));
         source1Interface->playBuffer((char*)&sine_330hz_16b_8k_signed[waveHeaderSize], 
                                      sizeof(sine_330hz_16b_8k_signed) - waveHeaderSize, 
                                      PROMPTBUF_SAMPLERATE,
@@ -2264,7 +2264,7 @@ class CpPhoneMediaInterfaceTest : public CppUnit::TestCase
 
         RTL_EVENT("Prompt", 1);
         printf("playing prompt in source 1\n");
-        CPPUNIT_ASSERT_EQUAL((unsigned)160116, sizeof(sine_330hz_16b_8k_signed));
+        CPPUNIT_ASSERT_EQUAL((size_t)160116, sizeof(sine_330hz_16b_8k_signed));
         source1Interface->playBuffer((char*)&sine_330hz_16b_8k_signed[waveHeaderSize], 
                                      sizeof(sine_330hz_16b_8k_signed) - waveHeaderSize, 
                                      PROMPTBUF_SAMPLERATE,
@@ -2318,7 +2318,7 @@ class CpPhoneMediaInterfaceTest : public CppUnit::TestCase
 
         RTL_EVENT("Prompt", 1);
         printf("playing prompt in source 1\n");
-        CPPUNIT_ASSERT_EQUAL((unsigned)160116, sizeof(sine_330hz_16b_8k_signed));
+        CPPUNIT_ASSERT_EQUAL((size_t)160116, sizeof(sine_330hz_16b_8k_signed));
         source1Interface->playBuffer((char*)&sine_330hz_16b_8k_signed[waveHeaderSize], 
                                      sizeof(sine_330hz_16b_8k_signed) - waveHeaderSize, 
                                      PROMPTBUF_SAMPLERATE,
@@ -2376,7 +2376,7 @@ class CpPhoneMediaInterfaceTest : public CppUnit::TestCase
 
         RTL_EVENT("Prompt", 1);
         printf("playing prompt in source 1\n");
-        CPPUNIT_ASSERT_EQUAL((unsigned)160116, sizeof(sine_330hz_16b_8k_signed));
+        CPPUNIT_ASSERT_EQUAL((size_t)160116, sizeof(sine_330hz_16b_8k_signed));
         source1Interface->playBuffer((char*)&sine_330hz_16b_8k_signed[waveHeaderSize], 
                                      sizeof(sine_330hz_16b_8k_signed) - waveHeaderSize, 
                                      PROMPTBUF_SAMPLERATE,
@@ -2434,7 +2434,7 @@ class CpPhoneMediaInterfaceTest : public CppUnit::TestCase
 
         RTL_EVENT("Prompt", 1);
         printf("playing prompt in source 1\n");
-        CPPUNIT_ASSERT_EQUAL((unsigned)160116, sizeof(sine_330hz_16b_8k_signed));
+        CPPUNIT_ASSERT_EQUAL((size_t)160116, sizeof(sine_330hz_16b_8k_signed));
         source1Interface->playBuffer((char*)&sine_330hz_16b_8k_signed[waveHeaderSize], 
                                      sizeof(sine_330hz_16b_8k_signed) - waveHeaderSize, 
                                      PROMPTBUF_SAMPLERATE,
